+++ fabric/src/main/kotlin/dev/nyon/magnetic/config/ConfigCache.kt
@@ -10,13 +10,14 @@ var ignoredEntities: Set<ResourceLocation> = setOf()
 
 private val registryAccess by lazy { RegistryAccess.ImmutableRegistryAccess(listOf(BuiltInRegistries.ENTITY_TYPE)) }
 private val registry by lazy { registryAccess.lookupOrThrow(Registries.ENTITY_TYPE) }
-internal fun loadIgnoredEntities(): Set<ResourceLocation> {
+
+internal fun reloadIgnoredEntities() {
     val ignored: MutableSet<ResourceLocation> = mutableSetOf()
     config.ignoreEntities.forEach { (original, isTag) ->
         if (!isTag) ignored.add(original)
         else ignored.addAll(original.getTagEntries())
     }
-    return ignored
+    ignoredEntities = ignored
 }
 
 private fun ResourceLocation.getTagEntries(): List<ResourceLocation> {