+++ fabric/src/main/kotlin/dev/nyon/magnetic/config/conditions/Statement.kt
@@ -1,0 +1,7 @@
+package dev.nyon.magnetic.config.conditions
+
+internal val statements: Set<Statement> = setOf(conditions, operators).flatten().toSet()
+
+sealed interface Statement {
+    val identifiers: Set<String>
+}